use new component for status and move code in pif store
