+++ @xen-orchestra/lite/src/stores/xen-api/pif.store.ts
@@ -1,9 +1,27 @@
+import type { Status } from '@/components/pool/network/PoolNetworksTable.vue'
+import type { XenApiPif } from '@/libs/xen-api/xen-api.types'
 import { createXapiStoreConfig } from '@/stores/xen-api/create-xapi-store-config'
 import { createSubscribableStoreContext } from '@core/utils/create-subscribable-store-context.util'
 import { defineStore } from 'pinia'
 
 export const usePifStore = defineStore('xen-api-pif', () => {
-  const config = createXapiStoreConfig('pif')
+  const { context: baseContext, ...configRest } = createXapiStoreConfig('pif')
 
-  return createSubscribableStoreContext(config, {})
+  const determineStatus = (PIFs: XenApiPif[]): Status => {
+    if (PIFs.length === 0) {
+      return 'disconnected'
+    }
+    const currentlyAttached = PIFs.map(PIF => PIF.currently_attached)
+    if (currentlyAttached.every(Boolean)) {
+      return 'connected'
+    }
+    if (currentlyAttached.some(Boolean)) {
+      return 'partially-connected'
+    }
+    return 'disconnected'
+  }
+
+  const context = { ...baseContext, determineStatus }
+
+  return createSubscribableStoreContext({ context, ...configRest }, {})
 })
