+++ @xen-orchestra/web-core/lib/components/ui/info/UiInfo.vue
@@ -44,6 +44,7 @@ const icon = computed(() => iconByAccent[accent])
 
 <style lang="postcss" scoped>
 .ui-info {
+  align-items: start;
   display: flex;
   gap: 0.8rem;
